Move validators out of Microsoft.Diagnostics.Monitoring.Options
Because they rely on .NET10+ ValidationContext ctor
