+++ src/Tools/dotnet-monitor/OptionsValidators/FileSystemEgressProviderOptionsValidator.cs
