+++ src/Tools/dotnet-monitor/OptionsValidators/GlobalProviderOptionsValidator.cs
@@ -1,0 +1,12 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using Microsoft.Extensions.Options;
+
+namespace Microsoft.Diagnostics.Monitoring.WebApi
+{
+    [OptionsValidator]
+    partial class GlobalProviderOptionsValidator : IValidateOptions<GlobalProviderOptions>
+    {
+    }
+}
