+++ src/Tools/dotnet-monitor/OptionsValidators/MetricsOptionsValidator.cs
